Hosten er sat op til at virker

+++ Client.Domain/Product.cs
@@ -8,7 +8,7 @@ namespace Client.Domain {
     public class Product {
         public int ID { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
         public int MinStock { get; set; }
         public int MaxStock { get; set; }
@@ -19,7 +19,7 @@ namespace Client.Domain {
         public List<Keyword> Keywords { get; set; }
         public Review Review { get; set; }
 
-        public Product(string name, double price, int stock, int minStock, int maxStock, string description, List<Copy> copies, Category category, List<Keyword> keywords) {
+        public Product(string name, decimal price, int stock, int minStock, int maxStock, string description, List<Copy> copies, Category category, List<Keyword> keywords) {
             Name = name;
             Price = price;
             Stock = stock;
