+++ ShoppingService/ProductService.cs
@@ -16,7 +16,7 @@ namespace Server.ServiceLayer
         public ProductService() {
             productDb = new ProductDB();
         }
-        public void CreateProduct(string name, double price, int stock, int minStock, int maxStock, string description) {
+        public void CreateProduct(string name, decimal price, int stock, int minStock, int maxStock, string description) {
             Product p = new Product(name, price, stock, minStock, maxStock, description);
             productDb.Create(p);
         }
